code refactoring to split out run_migrate
