+++ migration/build.rs
@@ -25,7 +25,6 @@ fn main() {
     }
 
     let mut code = String::new();
-    code.push_str("use std::sync::Arc;\n");
     code.push_str("pub fn add_migrations(db_migrate: &mut DbMigrate) {\n");
 
     for migration in migrations {
