+++ migration/src/main.rs
@@ -1,9 +1,13 @@
 use clap::Parser;
 use fiber::store::db_migrate::DbMigrate;
 use fnn_migrate::migrations::*;
+use fnn_migrate::util::prompt;
 use rocksdb::ops::Open;
 use rocksdb::{DBCompressionType, Options, DB};
 use std::path::Path;
+use std::process::exit;
+use std::{cmp::Ordering, sync::Arc};
+use tracing::error;
 use tracing_subscriber;
 
 include!(concat!(env!("OUT_DIR"), "/migrations.rs"));
@@ -34,6 +38,40 @@ struct Args {
     skip_confirm: bool,
 }
 
+fn run_migrate<P: AsRef<Path>>(
+    migrate: DbMigrate,
+    path: P,
+    skip_confirm: bool,
+) -> Result<Arc<DB>, String> {
+    if let Err(_) = migrate.init_or_check(path.as_ref()) {
+        let result = migrate.check();
+        if result == Ordering::Less {
+            if !skip_confirm {
+                let path_buf = path.as_ref().to_path_buf();
+                let input = prompt(format!("\
+                     Once the migration started, the data will be no longer compatible with all older version,\n\
+                     so we strongly recommended you to backup the old data {} before migrating.\n\
+                     \n\
+                     \nIf you want to migrate the data, please input YES, otherwise, the current process will exit.\n\
+                     > ", path_buf.display()).as_str());
+
+                if input.trim().to_lowercase() != "yes" {
+                    error!("Migration was declined since the user didn't confirm.");
+                    return Err("need to run database migration".to_string());
+                }
+            }
+            eprintln!("begin to migrate db ...");
+            let db = migrate.migrate().expect("failed to migrate db");
+            eprintln!("db migrated successfully, now your can restart the fiber node ...");
+            return Ok(db);
+        } else {
+            assert_eq!(result, Ordering::Greater);
+            return Err("incompatible database, need to upgrade fiber binary".to_string());
+        }
+    }
+    Ok(migrate.db())
+}
+
 fn main() {
     tracing_subscriber::fmt()
         .with_max_level(tracing::Level::DEBUG)
@@ -42,12 +80,12 @@ fn main() {
 
     let args = Args::parse();
     let path = Path::new(&args.path);
-
     let skip_confirm = args.skip_confirm;
 
     let db = open_db(path).expect("failed to open db");
     let migrate = init_db_migrate(db);
-    if let Err(err) = migrate.check_or_run_migrate(path, true, skip_confirm) {
+    if let Err(err) = run_migrate(migrate, path, skip_confirm) {
         eprintln!("{}", err);
+        exit(1);
     }
 }
